+++ apps/headless-crawler/lib/core/page/JobSearch/validators/error.ts
